requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
apscheduler==3.10.4
sqlite-utils==3.36
python-dotenv==1.0.1
openai==1.51.0
ics==0.7.2
